fix verify onSubmit verification

+++ src/containers/Auth/Verification/index.tsx
@@ -61,34 +61,40 @@ const Verification = ({isResetPassword, onSubmitResetPassword}:VerificationProps
     pageBackEvent()
     sendPhoneOTP()
    }, []);
-  const onSubmit = async (data: any) => {
-    if (userId) {
-      setIsVerifying(true);
-      const phoneVerificationResponse: any = await verifyPhoneOTP(data.code, userId);
-      if (phoneVerificationResponse?.data?.id) {
-        setIsVerifying(false);
-        toast.dismiss()
-        toast.success('Verified');
-        //localStorage.setItem('token', phoneVerificationResponse.token)
-        // process.env.REACT_APP_IS_BETA == 'TRUE' ? navigate('/') : navigate('/subscription');
-        navigate('/security');
-      }
-      else if (phoneVerificationResponse?.response?.status === 409) {
-        const phone = localStorage.getItem('phone');
-        // navigate('/password-reset',{state: {
-        //   username: phone,
-        //   code: data.code
-        // }}) 
-        localStorage.clear()
-        toast.error("User already exists")
-        navigate('/login')
-        // toast.error("It seems your phone number already registered in our system. Please try to login or recover your password.");
-        setIsVerifying(false);
-        // logoutClick();
-      } 
-      else if (phoneVerificationResponse?.response?.data) {
-        toast.info(phoneVerificationResponse?.response?.data?.details);
-        setIsVerifying(false);
+   const onSubmit = async (data: any) => {
+    if (isResetPassword) {
+      const code = getValues('code');
+      onSubmitResetPassword && onSubmitResetPassword({ code: code });
+    } else {
+      if (userId) {
+        setIsVerifying(true);
+        const phoneVerificationResponse: any = await verifyPhoneOTP(
+          data.code,
+          userId
+        );
+        if (phoneVerificationResponse?.data?.id) {
+          setIsVerifying(false);
+          toast.dismiss();
+          toast.success('Verified');
+          //localStorage.setItem('token', phoneVerificationResponse.token)
+          // process.env.REACT_APP_IS_BETA == 'TRUE' ? navigate('/') : navigate('/subscription');
+          navigate('/security');
+        } else if (phoneVerificationResponse?.response?.status === 409) {
+          const phone = localStorage.getItem('phone');
+          // navigate('/password-reset',{state: {
+          //   username: phone,
+          //   code: data.code
+          // }})
+          localStorage.clear();
+          toast.error('User already exists');
+          navigate('/login');
+          // toast.error("It seems your phone number already registered in our system. Please try to login or recover your password.");
+          setIsVerifying(false);
+          // logoutClick();
+        } else if (phoneVerificationResponse?.response?.data) {
+          toast.info(phoneVerificationResponse?.response?.data?.details);
+          setIsVerifying(false);
+        }
       }
     }
   };
@@ -181,16 +187,7 @@ const Verification = ({isResetPassword, onSubmitResetPassword}:VerificationProps
             open={errors.code ? true : false}
           />
           <Button
-            onClick={()=>{
-              if(isResetPassword)
-              {
-                const code=getValues("code")
-                onSubmitResetPassword && onSubmitResetPassword({code : code})
-              }
-              else{
-              handleSubmit(onSubmit)
-              }
-            }}
+            onClick={handleSubmit(onSubmit)}
             className="Submit-Button"
             loading={isVerifying}
             disabled={disableSubmit}
